fix(corpus-cli): deduplicate htmlToText and fix &#039; entity bug
fetch.ts had its own stripHtmlTags with an entity map missing &#039;,
so Wikimedia artist names containing apostrophes came through as
literal "&#039;" while html.ts decoded them correctly. Extracted the
html.ts variant (the complete one) into text.ts and deleted the
duplicate from fetch.ts.

Also named the magic 200 (max attribution length) and 80 (evidence
context length) in html.ts while in the neighbourhood.
